+++ src/pages/posts.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 
-import Redirect from "../components/redirect"
+import Redirect from "../components/Redirect"
 
 const PostsPage = (): JSX.Element => (
   <Redirect title="dev.to" url="https://dev.to/angelxehg" />
